+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the Audio Player project will be documented in this file.
 ## 3.7.0 - 2025-10-xx
 ### Changed
 - reduced playlist columns for streams
+- migrated controller annotations to PHP attributes
 
 ## 3.6.1 - 2025-10-24
 ### Changed